moved DefaultColorPicker.svelte
