+++ src/components/controls/DefaultColorPicker.svelte
