+++ src/components/controls/IntervalsLegendColorMapRow.svelte
@@ -3,7 +3,7 @@
   import { fade } from 'svelte/transition'
   import chroma from 'chroma-js'
   import { abs } from 'mathjs'
-  import DefaultColorPicker from '$components/DefaultColorPicker.svelte'
+  import DefaultColorPicker from '$components/controls/DefaultColorPicker.svelte'
   import Popper from '$lib/popper'
   import type { Color, IntervalLegendColorMapRow, Layer } from '$lib/types'
